Changes the DirectConnectionState to use the default configs for AdminClients (#209)
Fixes #207

Changes the `ClientConfiguration` so that the static methods do use the default configurations for admin clients, consumers, and producers.

Note that we need the static methods in `ClientConfiguration` because `DirectConnectionState` needs them (and `PlatformConnectionState` will, too), and `ConnectionState` objects are not beans (and therefore can’t inject the `ClientConfigurator` bean) and are created as normal objects in many of the tests. But we still want the `ClientConfigurator` bean methods to find the connections from the `ConnectionManager` bean and the clusters in the `ClusterCache` bean, while the static methods (that do the real work) require the `ConnectionState` object to be passed in as parameters.