MINOR: Automatic async refreshing of connection state did not send events
We recently added an asynchronous refreshing of connection states. Any changes in the connection state (e.g., from `ATTEMPTING` to `CONNECTED`, we would not fire the local events indicating the change in connectivity.

This fixes that and encapsulates the change in state to ensure the `StateChangedListener` on each `ConnectionState` object is called when the status changes.
* A CONNECTED internal event is fired when a connection’s state was previously not connected (see below) but becomes connected, or when it was previously connected and the state has changed in some way.
* A DISCONNECTED internal event is fired when a connection’s state was previously connected (see below) but becomes disconnected, or when it was not previously connected and the state has changed in some way.

This logic relies upon the `ConnectionState.isConnected()` method to determine whether the connection is connected or not connected.