fix: Support consuming records compressed with snappy
This change fixes the build of native executables so that users can consume records that were compressed with the snappy algorithm.
To this end, we must integrate the native libs for snappy into the native executable, slightly increasing its size (cc @shouples).

FWIW, I've identified the GraalVM resource configuration to add by
[collecting metadata with GraalVM's tracing agent](https://www.graalvm.org/latest/reference-manual/native-image/metadata/AutomaticMetadataCollection/).

Fixes #304.