Return meaningful error message when consuming records compressed with snappy failed (#342)
* fix: Returning meaningful error message when consuming records compressed with snappy failed

At the moment, the sidecar is not able to consume records that were compressed with snappy
when running in native mode. When attempting to do so, the REST API will return a response similar to the following one:

```
{
  "details": "Error id 88fab26a-4535-4a66-bf06-b037dac1fb5f-1",
  "stack": ""
}
```

Unfortunately, this is not very helpful to users. Since fixing the underlying bug, captured in https://github.com/confluentinc/ide-sidecar/issues/304,
will take longer than expected and has been moved out of the GA scope, this change makes sure we return a more meaningful
error message to users, hopefully helping them to figure out why the extension's Message Viewer can't show records.
The new response looks similar to the following one:

```
{
  "status": "500",
  "code": "proxy_error",
  "title": "At the moment, we don't support consuming records that were compressed with snappy. Please visit https://github.com/confluentinc/ide-sidecar/issues/304 to learn more.",
  "id": "d27faa2c-5e1c-4ef1-8562-7c3192ac8bc7",
  "errors": [
    {
      "code": "proxy_error",
      "title": "At the moment, we don't support consuming records that were compressed with snappy. Please visit https://github.com/confluentinc/ide-sidecar/issues/304 to learn more.",
      "detail": "At the moment, we don't support consuming records that were compressed with snappy. Please visit https://github.com/confluentinc/ide-sidecar/issues/304 to learn more."
    }
  ]
}
```

Fixes #341

* chore: Handle `KafkaException`s separately

* Add integration tests to verify we can consume compressed records (#344)

* test: Verify we can consume compressed records

This change adds integration tests so that we can verify that we can
consume records that were compressed with one of the supported
compression algorithms.

Fixes #303

* fix: Add key/value serializer

* test: Verify behavior when consuming records compressed with Snappy

* chore: Refine test