feat: Proof of concept for the Language Service proxy
This change adds an early POC for the WebSockets-based proxy for the Flink Language Service.
It's listening on the path `/flsp` and allows the extension to interact with the CCloud Flink Language Service
without having to manage authentication tokens or deal with reconnection attempts.

The proxy is designed as a simple pass-through proxy. When the extension connects to it, the proxy
will establish a WebSockets connection to the CCloud Flink Language Service and send an initial authentication message.
From this point on, it will pass all messages sent by the extension as they are to the Flink Language Service and vice versa.

When the connection between the proxy and the Flink Language Service gets closed, e.g., because the data plane token has expired,
the proxy will attempt up to `MAX_RECONNECT_ATTEMPTS` reconnection attempts before closing the connection with the extension.
Every message that was sent by the Flink Language Service and received by the proxy resets the reconnect attempts counter.