fix: Do not use `Path.of` when handling URIs
This change removes the usage of `Path.of` when handling URIs.
It's not meant to be used for URIs and seems to cause problems on Windows,
because Windows doesn't allow question marks in path names.
This change also adds a few log statements to the `UriUtil` so that we can better
debug potential future issues.

Fixes #412