feat: Make Kerberos config file configurable via Preferences API (#422)
* feat: Make Kerberos config file configurable via Preferences API

Allow users to configure the Kerberos config file via the Preferences API,
which will fix a bug that prevents users from authenticating with a Kafka cluster using Kerberos on macOS.

Note that this change doesn't validate if the Kerberos config file is configured
before attempting the authentication with the Kafka cluster using Kerberos. It's to be discussed if this
should be validated on the extension side.

Fixes #420