fix: Allow consuming records compressed with zstd using Direct connections (#490)
* fix: Allow consuming records compressed with zstd using Direct connections

This change fixes #489 and allows users to use Direct connections to consume records that were compressed with zstd compression.
To this end, this change introduces a custom loader class, `ZstdNativeLoader`, so that we can point the zstd-jni library to code-signed native library files.
It also updates our CI pipeline to perform the code-signing.

* chore: Extract loading of native library into util class

* chore: Add some basic tests

* chore: Include zstd-jni in THIRD_PARTY_NOTICES.txt