fix: Support passing `null` keys/values to the proxy of CCloud's Records API (#517)
* fix: Support passing `null` keys/values to the proxied CCloud Records API

This change fixes a potential `NullPointerException` in the proxy of Confluent Cloud's Records API
that is used to produce new records to a specific Kafka topic. The `NullPointerException` has been thrown
when a `null` value was provided as the key or the value of the to-be-produced record.

Fixes #516
Fixes https://github.com/confluentinc/vscode/issues/3065

* chore: Add JavaDocs

* chore: Add comments

* chore: Remove duplicate code from tests

* chore: Update comments

* chore: Remove unused import

* chore: Remove duplicate code